+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2025-07-10
 
+- [PR-33](https://github.com/itk-dev/tidy-feedback/pull/33)
+  - Fixed dynamic class in Svelte
+  - Cleaned up widget code
 - [PR-30](https://github.com/itk-dev/tidy-feedback/pull/30)
   29: Replaced Bootstrap with CoreUI
 - [PR-28](https://github.com/itk-dev/tidy-feedback/pull/28)